refactor: lottery 가 아닌 lotteryNumbers 를 갖도록 수정
