+++ src/main/kotlin/lottery/domain/WinningLottery.kt
@@ -1,9 +1,13 @@
 package lottery.domain
 
-data class WinningLottery(val lottery: Lottery) {
+data class WinningLottery(val lotteryNumbers: LotteryNumbers) {
+    fun countMatchedNumber(other: Lottery): Int {
+        return lotteryNumbers.countMatchedNumber(other.lotteryNumbers)
+    }
+
     companion object {
         fun create(numbers: List<Int>): WinningLottery {
-            return WinningLottery(Lottery(LotteryNumbers(numbers.toSet())))
+            return WinningLottery(LotteryNumbers(numbers.toSet()))
         }
     }
 }
