+++ src/test/kotlin/lottery/domain/LotteryNumbersTest.kt
@@ -2,6 +2,10 @@ package lottery.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
+import io.kotest.data.forAll
+import io.kotest.data.headers
+import io.kotest.data.row
+import io.kotest.data.table
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.ints.shouldBeInRange
 import io.kotest.matchers.shouldBe
@@ -29,6 +33,20 @@ class LotteryNumbersTest : StringSpec({
     "로또 번호가 1~45 범위를 벗어날 시 예외 발생한다" {
         listOf(0, -1, 46, 47).forAll { overNumber ->
             shouldThrow<IllegalArgumentException> { LotteryNumbers(setOf(overNumber, 1, 1, 1, 1, 1)) }
+        }
+    }
+
+    "로또번호끼리의 일치 수를 반환한다" {
+        table(
+            headers("lotteryNumbers", "otherNumbers", "expected"),
+            row(setOf(1, 2, 3, 4, 5, 6), setOf(1, 2, 3, 4, 5, 6), 6),
+            row(setOf(1, 2, 3, 4, 5, 6), setOf(1, 2, 3, 4, 5, 32), 5),
+            row(setOf(10, 11, 12, 13, 14, 15), setOf(15, 14, 13, 12, 40, 41), 4),
+            row(setOf(31, 32, 3, 4, 5, 6), setOf(31, 32, 40, 41, 42, 43), 2),
+        ).forAll { lotteryNumbers, others, expected ->
+            val origin = LotteryNumbers(lotteryNumbers)
+            val other = LotteryNumbers(others)
+            origin.countMatchedNumber(other) shouldBe expected
         }
     }
 })
